home/forms.py
from crispy_forms.helper import FormHelper
from crispy_forms.layout import Layout, Submit
from crispy_forms.layout import Submit
from django.forms import ModelForm, forms
from .models import Customer, Sender, newdestination

from django.forms import ModelForm
from django import forms


class CustomerForm(ModelForm):
    class Meta:
        model = Customer
        fields = '__all__'

        def __init__(self, *args, **kwargs):
            super().__init__(*args, **kwargs)

            self.helper - FormHelper
            self.helper.form_method = 'post'

class SenderForm(ModelForm):
    class Meta:
        model = Sender
        fields = '__all__'


class newdestinationForm(ModelForm):
    class Meta:
        model = newdestination
        fields = '__all__'

        def __init__(self, *args, **kwargs):
            super().__init__(*args, **kwargs)

            self.helper - FormHelper
            self.helper.form_method = 'post'
