requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-crispy-forms==1.14.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
